refactor: unit tests to main code structure
